wip: each addy has own comp
